Ajout d'un test pour éviter un warning lors du ménage.
git-svn-id: http://svn.sylogix.org/gepiphp/trunk@7373 e073496e-d129-0410-8f7b-841bfd467322

+++ cahier_texte_2/export_cdt.php
@@ -1292,7 +1292,8 @@ if($action=='export_zip') {
 	// On fait le ménage
 	for($i=0;$i<count($tab_fichiers_a_zipper);$i++) {
 		//echo "unlink($tab_fichiers_a_zipper[$i]);<br />";
-		unlink($tab_fichiers_a_zipper[$i]);
+		//unlink($tab_fichiers_a_zipper[$i]);
+		if(file_exists($tab_fichiers_a_zipper[$i])) {unlink($tab_fichiers_a_zipper[$i]);}
 	}
 	
 	rmdir($dossier_export."/cahier_texte");
